requirements.txt
streamlit>=1.40.0
pyyaml==6.0.3
python-docx==1.1.2
langgraph==0.2.40
pydantic>=2.9.0,<3
requests==2.32.5
google-genai==1.66.0